+++ dashboard/pages/About.py
@@ -1,3 +1,4 @@
+# pylint: disable=C0103
 """
 This is the "About" page of the dashboard, including information about the project and dashboard.
 """
